% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_posterior.R
\name{calc_posterior}
\alias{calc_posterior}
\title{Calculate a single posterior probability}
\usage{
calc_posterior(
  y,
  n,
  direction = "greater",
  p = NULL,
  delta = 0,
  prior = c(0.5, 0.5),
  S = 5000,
  seed = 1
)
}
\arguments{
\item{y}{vector of length two containing total responses c(y0, y1) for
two-sample case; integer of total responses for one-sample case}

\item{n}{vector of length two containing total sample size c(n0, n1) for
two-sample case; integer of total sample size for one-sample case}

\item{direction}{"greater" (default) if interest is in p(y1 > y0) and "less"
if interest is in p(y1 < y0). For one-sample case, "greater" if interest is
in p(y1 > p) and "less" if interest is in p(y1 < p)}

\item{delta}{clinically meaningful difference between groups.
Typically 0 (default).}

\item{prior}{hyperparameters of beta distribution, Beta(0.5, 0.5) is default}

\item{S}{number of samples, default is 5000}

\item{seed}{set the seed for the random number generation}
}
\value{
Returns the posterior probability of interest
}
\description{
This function is meant to be used in the context of a
clinical trial with a binary endpoint. Interest is in the response rate.
For the two-sample case, the total number of responses for the
standard-of-care arm is y0 and the total number of responses for the
experimental arm is y1. The function samples from the posterior beta
distribution based on the data and the prior beta hyperparameters, and
returns the posterior probability that y1 is greater than (or less than) y0
given the data. The one-sample case is also available, in which a target p
must be specified and the function returns the posterior probability that
y1 is greater than (or less than) p given the data.
}
\examples{

# Two-sample case
calc_posterior(y = c(15, 23), n = c(100, 100))

# One-sample case
calc_posterior(y = 24, n = 100, direction = "greater", p = 0.2, delta = NULL)

}
