% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample_cal_tbl-data.R
\docType{data}
\name{one_sample_cal_tbl}
\alias{one_sample_cal_tbl}
\title{Output from a one-sample call to \code{calibrate_thresholds}}
\format{
A list containing a
\enumerate{
\item a tibble 'res_summary' containing the posterior probability threshold
(pp_threshold); the predictive probability threshold (ppp_threshold);
the mean sample size under the null (mean_n1_null) and alternative
(mean_n1_alt) response rates; the proportion of positive trials under
the null (prop_pos_null) and alternative (prop_pos_alt) response rates; and
the proportion of trials stopped under the null (prop_stopped_null) and
alternative (prop_stopped_alt) response rates
\item 'call_list' containing the original function call
\item 'inputs' a list containing the inputs to the
original function call
}
}
\usage{
data(one_sample_cal_tbl)
}
\description{
This .rda file contains output from a one-sample call to
\code{calibrate_thresholds()}.
See the vignette titled "One-sample expansion cohort" for a description of
the input parameters used, or run
\code{one_sample_cal_tbl$inputs}
to see a list of the original function inputs. For use in testing
functions and in vignettes.
}
\keyword{datasets}
