% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample_decision_tbl-data.R
\docType{data}
\name{one_sample_decision_tbl}
\alias{one_sample_decision_tbl}
\title{Output from a one-sample call to \code{calc_decision_rules}}
\format{
A tibble containing n, the number of patients enrolled at each
futility monitoring point, and r, the number of responses at which we would
stop the trial at a given look if the number of observed responses is <=r,
or at the end of the trial the treatment is considered promising if the
number of observed responses is >r.
}
\usage{
data(one_sample_decision_tbl)
}
\description{
This .rda file contains output from a one-sample call to
\code{calc_decision_rules()}.
See the vignette titled "One-sample expansion cohort" for a description of
the input parameters used.
}
\keyword{datasets}
