% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_thresholds.R
\name{optimize}
\alias{optimize}
\title{Custom optimization method for \code{calibrate_thresholds} objects}
\usage{
optimize(x)
}
\arguments{
\item{x}{an object of class 'calibrate_thresholds', usually returned by the
\code{calibrate_thresholds} function}
}
\description{
Determines the optimal designs based on a variety of criteria
}
