% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_design.R
\name{optimize_design}
\alias{optimize_design}
\title{Function to setup usage of \code{optimize_design.calibrate_thresholds}}
\usage{
optimize_design(x, type1_range = c(0, 1), minimum_power = 0, ...)
}
\arguments{
\item{x}{an object of class 'calibrate_thresholds', usually returned by the
\code{calibrate_thresholds} function}

\item{type1_range}{a vector specifying the minimum and maximum acceptable
type I error. Specify NULL to return the full range of resulting type I
error. Defaults to c(0, 1) to return all results.}

\item{minimum_power}{a numeric between 0 and 1 specifying the minimum
acceptable power. Specify NULL to return the full range of resulting power.
Defaults to 0 to return all results.}

\item{...}{ignored}
}
\value{
No return value, called for side effects
}
\description{
Function to setup usage of \code{optimize_design.calibrate_thresholds}
}
