% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.calc_decision_rules}
\alias{plot.calc_decision_rules}
\title{Plot method for \code{calc_decision_rules} objects}
\usage{
\method{plot}{calc_decision_rules}(x, ...)
}
\arguments{
\item{x}{an object of class 'calc_decision_rules', usually returned by the
\code{calc_decision_rules} function}

\item{...}{unused}
}
\description{
Returns a plot of decision rules from the results of
\code{calc_decision_rules} that can interactively show when to stop and
when to proceed at the various interim analyses
}
