% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_thresholds.R
\name{plot.calibrate_thresholds}
\alias{plot.calibrate_thresholds}
\title{Plot method for \code{calibrate_thresholds} objects}
\usage{
\method{plot}{calibrate_thresholds}(x, type1_range = c(0.05, 0.1), minimum_power = 0.8, ...)
}
\arguments{
\item{x}{an object of class 'calibrate_thresholds', usually returned by the
\code{calibrate_thresholds} function}
}
\description{
Plot method for \code{calibrate_thresholds} objects
}
