% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppseq-package.R
\docType{package}
\name{ppseq-package}
\alias{ppseq}
\alias{ppseq-package}
\title{ppseq: Design a clinical trial using sequential predictive probablity monitoring}
\description{
This package provides functions to design clinical trials using sequential predictive probability monitoring. Functions are available to calibrate designs over a range of posterior and predictive thresholds, to plot the various design options, and to obtain the operating characteristics of optimal accuracy and optimal efficiency designs.
}
\author{
\strong{Maintainer}: Emily C. Zabor \email{zabore2@ccf.org} (\href{https://orcid.org/0000-0002-1402-4498}{ORCID})

Authors:
\itemize{
  \item Brian P. Hobbs
  \item Michael J. Kane
}

}
\keyword{internal}
