% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_thresholds.R
\name{print.calibrate_thresholds}
\alias{print.calibrate_thresholds}
\title{Print method for \code{calibrate_thresholds} objects}
\usage{
\method{print}{calibrate_thresholds}(x, ...)
}
\arguments{
\item{x}{an object of class 'calibrate_thresholds', usually returned by the
\code{calibrate_thresholds} function}
}
\value{
Returns a tibble
}
\description{
By default prints only the res_summary table from an object of
class 'calibrate_thresholds'
}
