% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{temp}
\alias{temp}
\title{Temporary R file to push the R folder to GitHub}
\usage{
temp(x, y)
}
\arguments{
\item{x}{a numeric value}

\item{y}{a numeric value}
}
\value{
Returns the sum of \code{x} and \code{y}
}
\description{
This function doesn't do anything useful, and I just created it
so that the R folder would be pushed to GitHub as empty folders don't get
committed. Will delete once we add real functions.
}
\examples{

temp(1, 10)
temp(2, NA)

}
