% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_decision_rules.R
\name{calc_decision_rules}
\alias{calc_decision_rules}
\title{Calculate a decision rule table for interim monitoring of a pre-specified
design}
\usage{
calc_decision_rules(
  n,
  N,
  theta,
  ppp,
  p0,
  direction = "greater",
  delta = NULL,
  prior = c(0.5, 0.5),
  S = 5000
)
}
\arguments{
\item{n}{integer of sample size so far for one-sample case}

\item{N}{tinteger of total planned sample size at end of trial N
for one-sample case}

\item{theta}{The target posterior probability. e.g. Efficacy decision if
P(p > p0) > theta for the one-sample case with greater direction.}

\item{ppp}{The target predictive probability. e.g. Stop the trial if the
predictive probability falls below this target.}

\item{p0}{The target value to compare to in the one-sample case
(i.e. the null response rate)}

\item{direction}{For one-sample case, "greater" if interest is in P(p > p0)
and "less" if interest is in P(p < p0).}

\item{delta}{clinically meaningful difference between groups.
Typically 0 for two-sample case. NULL for one-sample case (default).}

\item{prior}{hyperparameters of prior beta distribution.
Beta(0.5, 0.5) is default}

\item{S}{number of samples, default is 5000}
}
\value{
Returns a tibble with n at each look and r. Stop the trial at that
look if the number of observed responses is <=r. At the end of the trial,
the treatment is considered promising if the number of observed responses
is >r.
}
\description{
This function will take the posterior and predictive thresholds
for a pre-specified design with a given null response rate and fixed interim
looks and total sample size, and return the decision rules at each interim
analysis and the end of the trial. Intended for use after selecting an
optimal design using the functions \code{calibrate_thresholds} and
\code{optimize_design}.
}
\examples{
\dontrun{
set.seed(123)

# One-sample case
calc_decision_rules(seq(5, 25, 5), p0 = 0.1, N = 25, theta = 0.86, ppp = 0.2)
}

}
