% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_thresholds.R
\name{calibrate_thresholds}
\alias{calibrate_thresholds}
\title{Calibrate according to posterior probability threshold and predictive
probability threshold with interim futility monitoring}
\usage{
calibrate_thresholds(
  prob_null,
  prob_alt,
  n,
  direction = "greater",
  p0 = NULL,
  delta = 0,
  prior = c(0.5, 0.5),
  S = 5000,
  N,
  nsim = 1000,
  pp_threshold,
  ppp_threshold
)
}
\arguments{
\item{prob_null}{vector of length two containing the probability of event in
the standard of care and experimental arm c(p0, p1) for the two-sample case
for the null scenario;
integer of event probability for one-sample case}

\item{prob_alt}{vector of length two containing the probability of event in
the standard of care and experimental arm c(p0, p1) for the two-sample case
for the alternative scenario;
integer of event probability for one-sample case}

\item{n}{matrix containing the total number of patients accrued so far at
each interim look in the standard of care (column 1) and experimental
(column 2) arms for two-sample case; vector of sample size accrued so far
at each interim look for one-sample case. The last value should be equal to
the total sample size at the end of the trial.
If only a single look will be done
at the end of the trial, this can be a vector specifying the total sample
size c(N0, N1) for the two-sample case or an integer specifying the total
sample size N for the one-sample case}

\item{direction}{"greater" (default) if interest is in p(p1 > p0) and "less"
if interest is in p(p1 < p0) for two-sample case. For one-sample case,
"greater" if interest is in p(p > p0) and "less" if interest is in p(p < p0).}

\item{p0}{The target value to compare to in the one-sample case}

\item{delta}{clinically meaningful difference between groups.
Typically 0 (default).}

\item{prior}{hyperparameters of prior beta distribution.
Beta(0.5, 0.5) is default}

\item{S}{number of samples drawn from the posterior. Default is 5000}

\item{N}{the total planned sample size at the end of the trial, c(N0, N1)
for two-sample case; integer of total planned sample size at end of trial N
for one-sample case}

\item{nsim}{Number of simulated trial datasets.}

\item{pp_threshold}{the posterior probability threshold of interest}

\item{ppp_threshold}{the posterior probability threshold of interest for
futility monitoring}
}
\value{
A list containing a
\enumerate{
\item a tibble 'res_summary' containing the posterior probability threshold
(pp_threshold), the predictive probability threshold (ppp_threshold),
the mean sample size (mean_n0 and mean_n1 for two-sample case; mean_n1 for
one-sample case), and the proportion of positive trials.
\item 'call_list' containing the original function call
\item 'calibrate_thresholds_inputs' a list containing the inputs to the
original function call
}

The proportion of positive trials will be a measure of the type I error
for a null setting, and a measure of the power in the alternative setting.
}
\description{
This function is meant to be used in the context of a
clinical trial with a binary endpoint. For a vector of possible posterior
decision thresholds, the function simulates many trials and then calculates
the average number of times the posterior probability exceeds a given
threshold. In a null case, this will result in the type I error at a given
threshold. In an alternative case, this will result in the power at a given
threshold.
}
\examples{

# One-sample case
\dontrun{
set.seed(123)

calibrate_thresholds(prob_null = 0.1, prob_alt = 0.3, 
n = seq(5, 25, 5), direction = "greater", p0 = 0.1, delta = NULL, 
prior = c(0.5, 0.5), S = 5000, N = 25, nsim = 1000, 
pp_threshold = c(0.9, 0.95, 0.96, 0.98), 
ppp_threshold = seq(0.05, 0.2, 0.05))
}

# Two-sample case

}
