% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.calibrate_thresholds}
\alias{plot.calibrate_thresholds}
\title{Plot method for \code{calibrate_thresholds} objects}
\usage{
\method{plot}{calibrate_thresholds}(x, type1_range = c(0.05, 0.1), minimum_power = 0.8, plotly = FALSE, ...)
}
\arguments{
\item{x}{an object of class 'calibrate_thresholds', usually returned by the
\code{calibrate_thresholds} function}

\item{type1_range}{a vector specifying the minimum and maximum acceptable
type I error. Specify c(0, 1) to return the full range of resulting type I
error. Defaults to c(0.05, 0.1)}

\item{minimum_power}{a numeric between 0 and 1 specifying the minimum
acceptable power. Specify 0 to return the full range of resulting power.
Defaults to 0.8.}

\item{plotly}{a logical indicator of whether you want the plots returned as
interactive plotly plots or non-interactive ggplots. Defaults to FALSE.}

\item{...}{unused}
}
\description{
Returns two interactive \code{plotly} plots to compare results
from various designs generated from a call to \code{calibrate_thresholds}
based on various criteria, and to assist in selecting an optimal design.
}
