% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppseq-package.R
\docType{package}
\name{ppseq-package}
\alias{ppseq}
\alias{ppseq-package}
\title{ppseq: Conduct sequential posterior probablity monitoring}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Emily C. Zabor \email{zabore2@ccf.org} (\href{https://orcid.org/0000-0002-1402-4498}{ORCID})

Authors:
\itemize{
  \item Brian P. Hobbs
  \item Alex Kaizer
}

}
\keyword{internal}
